Fix breakage after EABI fix patch
The 'a' flag on the archive section got lost, meaning the whole
content of the built CPIO blob was put into a section that was
discarded on load.  It turns out that the Sabre (at least) also needs
write access to the CPIO blob.

Also, IA32 and X86_64 use the same compiler; use the -m32 flag to
force 32-bit ELF output for IA32.