add CMake level memoization of build rules
Add memoize_add_custom_command function which is intended to replace
complex add_custom_command calls and provide a memoization feature where
the results of the rule are stored in an external cache. Future builds
will be able to reuse the artifacts from the cache if the command can
infer that the inpur arguments are the same.

SEL4_CACHE_DIR can either be set in the environment or passed in to
CMake with -DSEL4_CACHE_DIR=path-to-cache-dir. IF SEL4_CACHE_DIR is not
set then the cache won't be used and this feature should have no effect.