cmake,memoize: Check empty path before absolute
If SEL4_CACHE_DIR wasn't set then a check to IS_ABSOLUTE would return
false, causing an attempted conversion from a relative path to absolute
path and set the cache dir to the top of the build directory. Checking
if the path is empty initially and then avoiding the conversion, leaves
the string empty.