hifive: Fixed UART Get Character function
The previous implementation made two reads to the rxdata register: One
to check if the Data Empty flag was set and another to grab the actual
data. However, the HiFive UART's rxdata is cleared on read. So this
would return 0xff every time. Now the rxdata is read once, stored in a
variable and operations are done against the value stored in that
variable.

Change-Id: I8f6d746293d0ce39dc5dd42eca485a944f1383a2