elfloader: move self to correct address on aarch64
The "shoehorn" tool figures out an address that the elfloader
can be linked against based on a number of parameters known at
buildtime. Unfortunately, binary (as opposed to ELF or EFI) images
have no way of communicating this assumption to the bootloader,
so the bootloader usually loads us in the wrong place. Fix this
by making the elfloader move itself (if possible) to the
correct location.