Don't attempt to build hosted programs to test GCC
By building a program which provides `main()`, we were assuming that
there is a libc available to compiler. This is not the case for the Red
Hat provided cross-compilers which are intended for use in freestanding
environments (the example given is kernel code). Given that the seL4
build system assumes a freestanding environment it is sensible to only
check that the compiler being used is capable of building such programs.

This change alters the code built when testing compiler options in
`environment_flags.cmake`, and also adds fixed options to force the
compiler to not link against the stdlib or include crt objects. By doing
so we avoid any requirement for a target libc on the build host.