simulate.py: delay terminal reset on failure
A terminal reset clears the screen for xterm-style terminals.  If QEMU
fails, for instance because the platform does not support simulation,
its diagnostic messages will vanish before the user can read them.
(Modern versions of xterm clear the scrollback buffer on reset too, so
this is even more frustrating.)

Check QEMU's exit status, and if it is nonzero, sleep for 5 seconds
before resetting the terminal.