add_fpu_compilation_options: More accurate checks
The FPU checks need to check that the compiler can compile programs with
the desired FPU options AND link against compiler runtime libraries with
the correct FPU ABI. If either of these conditions aren't met, then we
can't build programs correctly.

We use the standard check_c_source_runs CMake function which inherits
the compilation options that have been set already. Unfortunately it
doesn't inherit any changes to CMAKE_<LANG>_LINK_EXECUTABLE variables
which is what we use to set the compiler libraries and so we need to
specify those again manually. This both compiles and links our test
program.  In cross-compiling environments, the run stage will be
skipped; yet this is still sufficient to pick up ABI mismatches.