simulate.py: delay for a short time on success
Per discussion with Kent, a brief delay even if QEMU exits with a
successful status may be desirable, because there may have been a
failure in the simulated environment.

We have a shorter delay in part because we presume that the screen will
be more deciperhable in the QEMU-successful case, because QEMU will put
the terminal into a sensible state itself before terminating.
Experience will tell us how reliable this assumption is.

Tested success case with:
    ../griddle --SIMULATION=ON --PLATFORM=sabre
    ./simulate
Tested failure case with:
    ./simulate --extra-qemu-args bogus