griddle: Use bash and explain why.
This script uses bash for its execution, but not because it uses any
Bashisms; instead, it is to work around a POSIX-violating bug in signal
handling in dash, which is the /bin/sh on most Debian-based systems.
See <https://bugs.debian.org/bug=683671>, filed in 2012 and still not
fixed (posh has the same problem; ksh, mksh, yash, and zsh do not).