pico_dev_eth: Bug fix
A simple linked-list bug: `buf_pool` has a free list for bookkeeping
free ethernet buffers, it keeps track of the index of the latest free
buffer (the head). Every time a new free buffer arrived, it updates
the head of the free list to the index of the new buffer, then stores
the index of the old head inside the new head's buffer. It wasn't
implemented properly as it stores the old head's index inside its own
buffer and does nothing to the new head's buffer.
This could cause random memory corruption bugs.

Fix: Store the index of the old head into the new free buffer correctly.