libethdrivers,tx2: Return bits in eqos_handle_irq
Now that the TX interrupt is also working, there may be a situation
where both RX and TX interrupts to fire at the same causing the common
interrupt line to mask either one of the two. This commit updates the
interrupt handler so that it returns a bitfield of interrupts seen so
that it avoids this situation.

Also replace the magic numbers in the driver with constants and make a
minor aesthetics change to make it more readable.