libethdrivers,tx2: Fix up descriptors and buffers
The address of the last descriptor was off by one and some descriptors
also had invalid buffers when refilling buffers for the RX ring. This
wasn't a problem for applications working under lower throughputs but on
higher throughputs, the Ethernet device would sometimes hang possibly
due to bus errors on the DMA controller. This commit fixes those issues.