libethdrivers,tx2: Refactor RX IRQ handling code
This commits refactors the RX path IRQ handling code so the process of
disabling RX interrupts is outside of the IRQ status checking code. It
makes more sense to disable interrupts in the IRQ handling function and
not in the IRQ status checking code. Also rename one of fields in the
DMA registers struct so that it matches the actual functionality.