libethdrivers,tx2: Fit TX,RX tail descriptors
The tail descriptors indicate the last descriptor that is available to
the DMA controller for the Ethernet module. While it is possible if we
set the tail descriptor to be the last descriptor in the allocated ring,
this could cause slowdowns when the DMA controller looks at a descriptor
in the ring that is not ready to be transmitted/received into.

This commit changes it so that the tail descriptors are set when we have
buffers ready and so also that it matches the last descriptor that is
available to be acted on.