elfloader: Make switch_to_hyp_mode more precise
Switching processor modes causes the sp and lr registers to be swapped
with the banked variants at the new mode. The switching code needs to
save the old values in registers that aren't banked while making the
switch. Doing this all within the same assembly block ensures that the
registers aren't placed placed on the stack by the compiler which it is
allowed to do.

Signed-off-by: Kent McLeod <Kent.Mcleod@data61.csiro.au>