libplatsupport,CMake: Check platsupport helpers
This commit adds a check to see if the path to the helpers is not
NULL/empty before including the helpers. As the comment in the commit
mentions, this is a hack to allow builds to be green before a more
suitable fix can be made to the older projects.