liblwip: Exclude lwipopts.h
Including the library's version of lwipopts.h config prevents users from
using a custom version of the lwipopts.h config file for each
application and instead restricts the user to having to modify the
library's version of lwipopts.h which isn't a suitable workflow.

This commit excludes the library's version of lwipopts.h from the build.
Users are now required to use the CMake macro inside the
lwip_helpers.cmake file to include a custom version of lwipopts.h. Also
update style filter to point to the new default lwipopts.h location.