risc-v: switch to hart1
seL4 and librarys assume booting on hart 1 while
Opensbi and U-boot pick a hart at random.

This patch uses the HSM SBI call (if it exisits)
to start the elfloader on hart1 if a different hart
has been chosen.

If SMP is enabled it now starts all harts with the HSM extension
from boot.c.

Credits go to Peter Chubb for initial spike.

Signed-off-by: Oliver Scott <Oliver.Scott@data61.csiro.au>