cmake-tool: move rpi3 to binary_list
The Raspberry Pi3 has its own definition for configuring the
setting for binary images builds for aarch64 only, meaning
it cannot be part of the more generic binary_list definition.

The 32-bit aarch32 rpi3 builds can also boot from a binary
image so there is no actual need for this sepcial handling.

Remove it and move the bcm2837/rpi3 match to the binary_list
definition instead.

Signed-off-by: Viktor Sannum <sannum.viktor@gmail.com>