elfloader: port fixup_image_base to C
Porting from assembly to a higher level language makes this function
more portable.

The assembly already called out to C functions implementing memmove, so
this shouldn't be introducing any new C runtime dependencies.

The assembly version had a weaker check for overlap that didn't
guarantee that the stack memory was protected during the relocation.
The C version just aborts if any part of the image is overlapping.
This is based on an assumption that the image is either going to be
loaded to the correct address, or to an incorrect address far enough
away from the correct address that there's no overlap.

The assembly version also called abort which tries to print before
looping, but print() isn't initialized yet and what's more, if the bss
hasn't been reset, then when print tries to call a putchar function
pointer, control flow can be hijacked. So instead we loop.

Signed-off-by: Kent McLeod <kent@kry10.com>