cmake_tool: add zynqmp to platform_strings
Without adding the zynqmp subplatforms (zcu102, ultra96, ultra96v2),
sel4test can only be compiled for the generic zynqmp platform. While
this is compatible with the ZCU102, the Ultra96 platforms use a
different serial port, so the sel4test logs would be unreadable.

This commit allows sel4test to be compiled with -DPLATFORM=ultra96
and -DPLATFORM=ultra96v2.

Signed-off-by: Chris Guikema <chris.guikema@dornerworks.com>