Shift 32-bit image start address on imx8mm
The port of the hardware debug API to aarch64 has increased the size
of some debug images to the point that the address range the image is
loaded at overlaps with the address range the elf-loader relocates to
and expects to run at. This was solved on aarch64 by shifting the uboot
load address to 0x42000000. We do not differentiate between aarch32 and
aarch64 images in uboot, and the aarch32 elfloader is not relocated, so
the configuration in this commit must be changed to have the aarch32
elfloader expect to be executed from the new load address.

Signed-off-by: Alwin Joshy <joshyalwin@gmail.com>