i.MX 93: Add Low Power Timer (LPTMR) Driver
Only used if the Arm generic timer cannot be used.

The clock source is configurable. The default is the 32.768 kHz clock,
as that is more accurate for timekeeping and keeps running when the
board is in low power state and can wake up the CPU.

Retrieving the time takes about 0.5 us, so this timer is very slow.

Both LPMTR1 and LPTMR2 are used because you are not allowed to
change the compare value while the timer is running.

Signed-off-by: Indan Zupancic <indan@nul.nu>