ltimer: support old clang without __attr__(error)
Clang didn't have `__attribute__((error(msg)))` until version 14,
but we support older versions of Clang. Both GCC and clang support
`__has_attribute(error)`; GCC since v5 (2015) and Clang seemingly
always.

This fixes builds that were failing due to 'unknown attribute error'.
We are not relying on the presence of this attribute for correct
output, but purely as a debug message which fails before the linker.
If we don't support the message failing at link-time is OK too.

Signed-off-by: julia <git.ts@trainwit.ch>