RISCV, Elfloader: Assume FIRST_HART_ID is first
The old code always skipped starting hart 0, which happened
to work because if FIRST_HART_ID is 0, we will skip doing
anything anyway. Instead of this subtle initialisation that
happens to work, make it explicit behaviour depending on
CONFIG_FIRST_HART_ID.

It seems that hart ID 0 is a special management core on the
HiFive, trying to start it either faults or halts the system.
So it relied on skipping 0 implicitly. Polarfire and Star64
are other platforms with non-zero first hart ID.

Signed-off-by: Indan Zupancic <indan@nul.nu>