libplatsupport: fix FDT translation of 'reg'
The Device Tree nodes for devices on certain platforms does not describe
the actual MMIO address in the 'reg' field. Instead, it can be an offset
of the parent bus address.

Rather than taking the values of 'reg' as the MMIO registers directly,
we have to check that the 'ranges' field exists on the parent and then
use it to potentially translate the child node's address.

I ran into this when porting the Raspberry Pi 5 to sel4test, where
the timer peripheral is allocated via the DTB node unlike other platform
support where the device address/IRQs are hard-coded.

See section 2.3.8 of v0.4 of the Device Tree specification for more
details.

Signed-off-by: Ivan Velickovic <i.velickovic@unsw.edu.au>